+++ packages/core/e2e/local-build.test.ts
@@ -14,6 +14,7 @@ describe.each([
   'nuxt',
   'hono',
   'express',
+  'astro',
 ])('e2e', (project) => {
   test('builds without errors', { timeout: 180_000 }, async () => {
     // skip if we're targeting specific app to test